Split segment and exnode declarations
For some reason, ex3.h contained a lot of segment declarations. Split
them apart
