+++ src/lio/bin/ex_get.c
@@ -19,6 +19,7 @@
 #include <apr_time.h>
 #include <gop/gop.h>
 #include <gop/types.h>
+#include <lio/segment.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
