+++ src/lio/bin/ex_inspect.c
@@ -18,6 +18,7 @@
 
 #include <gop/gop.h>
 #include <gop/types.h>
+#include <lio/segment.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
