+++ src/lio/lio_core.c
@@ -21,6 +21,7 @@
 #include <gop/opque.h>
 #include <gop/tp.h>
 #include <gop/types.h>
+#include <lio/segment.h>
 #include <stdio.h>
 #include <stdio.h>
 #include <stdlib.h>
