+++ src/lio/lio_core_os.c
@@ -22,6 +22,7 @@
 #include <gop/opque.h>
 #include <gop/tp.h>
 #include <gop/types.h>
+#include <lio/segment.h>
 #include <stdio.h>
 #include <stdio.h>
 #include <stdlib.h>
